requirements.txt
numpy >= 1.1
Pillow
matplotlib